+++ Language/Haskell/TH/Desugar/Sweeten.hs
@@ -217,24 +217,25 @@ letDecToTH (DPragmaD prag)      = fmap PragmaD (pragmaToTH prag)
 
 conToTH :: DCon -> Con
 #if __GLASGOW_HASKELL__ > 710
-conToTH (DCon [] [] n (DNormalC stys) (Just rty)) =
+conToTH (DCon [] [] n (DNormalC _ stys) (Just rty)) =
   GadtC [n] (map (second typeToTH) stys) (typeToTH rty)
-conToTH (DCon [] [] n (DInfixC sty1 sty2) (Just rty)) =
-  GadtC [n] (map (second typeToTH) [sty1,sty2]) (typeToTH rty)
 conToTH (DCon [] [] n (DRecC vstys) (Just rty)) =
   RecGadtC [n] (map (thirdOf3 typeToTH) vstys) (typeToTH rty)
 #endif
-conToTH (DCon [] [] n (DNormalC stys) _) =
-#if __GLASGOW_HASKELL__ > 710
-  NormalC n (map (second typeToTH) stys)
-#else
-  NormalC n (map (bangToStrict *** typeToTH) stys)
-#endif
-conToTH (DCon [] [] n (DInfixC sty1 sty2) _) =
+conToTH (DCon [] [] n (DNormalC True [sty1, sty2]) _) =
 #if __GLASGOW_HASKELL__ > 710
   InfixC (second typeToTH sty1) n (second typeToTH sty2)
 #else
   InfixC ((bangToStrict *** typeToTH) sty1) n ((bangToStrict *** typeToTH) sty2)
+#endif
+-- Note: it's possible that someone could pass in a DNormalC value that
+-- erroneously claims that it's declared infix (e.g., if has more than two
+-- fields), but we will fall back on NormalC in such a scenario.
+conToTH (DCon [] [] n (DNormalC _ stys) _) =
+#if __GLASGOW_HASKELL__ > 710
+  NormalC n (map (second typeToTH) stys)
+#else
+  NormalC n (map (bangToStrict *** typeToTH) stys)
 #endif
 conToTH (DCon [] [] n (DRecC vstys) _) =
 #if __GLASGOW_HASKELL__ > 710